+++ addons/web_editor/static/src/js/editor/smooth_scroll_on_drag.js
